requirements.txt
-e ./causal-conv1d
-e ./mamba
beartype==0.21.0
einops==0.8.1
h5py==3.10.0
matplotlib==3.7.2
numpy==1.26.4
opencv-python==4.8.1.78
Pillow==10.0.1
PyWavelets==1.4.1
scipy==1.11.3
shortuuid==1.0.13
tensorboardX==2.6.2.2
timm==0.9.7
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.4
